+++ data_pipelines_cli/cli_commands/compile.py
@@ -12,15 +12,15 @@ from ..cli_constants import (
     IMAGE_TAG_TO_REPLACE,
     INGEST_ENDPOINT_TO_REPLACE,
 )
-from ..cli_utils import (
-    echo_error,
-    echo_info,
-    echo_subinfo,
-    get_argument_or_environment_variable,
+from ..cli_utils import echo_error, echo_info, get_argument_or_environment_variable
+from ..config_generation import (
+    copy_config_dir_to_build_dir,
+    copy_dag_dir_to_build_dir,
+    generate_profiles_yml,
 )
 from ..data_structures import DockerArgs
+from ..dbt_utils import run_dbt_command
 from ..io_utils import replace
-from .dbt import dbt
 
 
 def _replace_image_tag(k8s_config: pathlib.Path, docker_args: DockerArgs) -> None:
@@ -38,7 +38,7 @@ def _replace_docker_repository_url(
     replace(k8s_config, DOCKER_REPOSITORY_URL_TO_REPLACE, docker_args.repository)
 
 
-def _docker_build(docker_args: DockerArgs):
+def _docker_build(docker_args: DockerArgs) -> None:
     try:
         import docker
     except ModuleNotFoundError:
@@ -62,19 +62,12 @@ def _docker_build(docker_args: DockerArgs):
 
 
 def _dbt_compile(env: str) -> None:
+    profiles_path = generate_profiles_yml(env)
     echo_info("Running dbt commands:")
-
-    echo_subinfo("dbt deps")
-    dbt(("deps",), env)
-
-    echo_subinfo("dbt compile")
-    dbt(("compile",), env)
-
-    echo_subinfo("dbt docs generate")
-    dbt(("docs", "generate"), env)
-
-    echo_subinfo("dbt source freshness")
-    dbt(("source", "freshness"), env)
+    run_dbt_command(("deps",), env, profiles_path)
+    run_dbt_command(("compile",), env, profiles_path)
+    run_dbt_command(("docs", "generate"), env, profiles_path)
+    run_dbt_command(("source", "freshness"), env, profiles_path)
 
 
 def _copy_dbt_manifest() -> None:
@@ -94,37 +87,19 @@ def _replace_datahub_address(datahub_address: str) -> None:
     )
 
 
-def _copy_src_dir_to_dst_dir(src_dir: pathlib.Path, dst_dir: pathlib.Path) -> None:
-    # It has to be deleted before copying, as `copytree` complains with
-    # `FileExistsError`
-    if pathlib.Path.exists(dst_dir):
-        shutil.rmtree(dst_dir)
-    shutil.copytree(src_dir, dst_dir)
-
-
-def _copy_dag_dir_to_build_dir() -> None:
-    dag_src_path = pathlib.Path.cwd().joinpath("dag")
-    dag_dst_path = BUILD_DIR.joinpath("dag")
-    _copy_src_dir_to_dst_dir(dag_src_path, dag_dst_path)
-
-
-def _copy_config_dir_to_build_dir() -> None:
-    config_src_path = pathlib.Path.cwd().joinpath("config")
-    dag_dst_path = BUILD_DIR.joinpath("dag", "config")
-    echo_info(f"Copying 'config' directory to {dag_dst_path}")
-    _copy_src_dir_to_dst_dir(config_src_path, dag_dst_path)
-
-
 def compile_project(
-    repository: Optional[str], datahub: Optional[str], docker_build: bool, env: str
+    repository: Optional[str],
+    datahub: Optional[str],
+    docker_build: bool,
+    env: str,
 ) -> None:
     datahub_address = get_argument_or_environment_variable(
         datahub, "datahub", DATAHUB_URL_ENV
     )
     docker_args = DockerArgs(repository)
 
-    _copy_dag_dir_to_build_dir()
-    _copy_config_dir_to_build_dir()
+    copy_dag_dir_to_build_dir()
+    copy_config_dir_to_build_dir()
 
     k8s_config: pathlib.Path = BUILD_DIR.joinpath("dag", "config", "base", "k8s.yml")
     _replace_image_tag(k8s_config, docker_args)
@@ -139,11 +114,14 @@ def compile_project(
 
 
 @click.command(name="compile")
+@click.option("--env", required=True)
 @click.option("--repository", default=None)
 @click.option("--datahub", default=None)
 @click.option("--docker-build", is_flag=True, default=False)
-@click.option("--env", default="gitlab")
 def compile_project_command(
-    repository: Optional[str], datahub: Optional[str], docker_build: bool, env: str
+    env: str,
+    repository: Optional[str],
+    datahub: Optional[str],
+    docker_build: bool,
 ) -> None:
     compile_project(repository, datahub, docker_build, env)