+++ data_pipelines_cli/cli_commands/deploy.py
@@ -2,15 +2,14 @@ import io
 import json
 import os
 import pathlib
-import subprocess
 import sys
-from typing import Dict, Optional
+from typing import Dict, Optional, cast
 
 import click
 import yaml
 
 from ..cli_constants import BUILD_DIR
-from ..cli_utils import echo_error, echo_info
+from ..cli_utils import echo_error, echo_info, subprocess_run
 from ..data_structures import DockerArgs
 from ..filesystem_utils import LocalRemoteSync
 
@@ -46,12 +45,12 @@ class DeployCommand:
         self._sync_bucket()
 
     @staticmethod
-    def _get_project_name():
+    def _get_project_name() -> str:
         with open(pathlib.Path().joinpath("dbt_project.yml")) as f:
             dbt_project_config = yaml.safe_load(f)
             return dbt_project_config["name"]
 
-    def _docker_push(self):
+    def _docker_push(self) -> None:
         try:
             import docker
         except ModuleNotFoundError:
@@ -62,15 +61,16 @@ class DeployCommand:
 
         echo_info("Pushing Docker image")
         docker_client = docker.from_env()
+        docker_args = cast(DockerArgs, self.docker_args)
         for line in docker_client.images.push(
-            repository=self.docker_args.repository,
-            tag=self.docker_args.commit_sha,
+            repository=docker_args.repository,
+            tag=docker_args.commit_sha,
             stream=True,
         ):
             click.echo(line, nl=False)
 
     @staticmethod
-    def _datahub_ingest():
+    def _datahub_ingest() -> None:
         try:
             import datahub  # noqa: F401
         except ModuleNotFoundError:
@@ -80,7 +80,7 @@ class DeployCommand:
             sys.exit(1)
 
         echo_info("Ingesting datahub metadata")
-        subprocess.run(
+        subprocess_run(
             [
                 "datahub",
                 "ingest",