+++ data_pipelines_cli/data_structures.py
@@ -3,7 +3,6 @@ from typing import Dict, Optional
 
 import yaml
 
-from data_pipelines_cli.cli_constants import CONFIGURATION_PATH
 from data_pipelines_cli.cli_utils import (
     echo_error,
     get_argument_or_environment_variable,
@@ -22,11 +21,14 @@ class TemplateConfig(TypedDict):
 
 
 class DataPipelinesConfig(TypedDict):
-    username: str
     templates: Dict[str, TemplateConfig]
+    vars: Dict[str, str]
 
 
 def read_config_or_exit() -> DataPipelinesConfig:
+    # Avoiding a dependency loop between `cli_constants` and `data_structures`
+    from data_pipelines_cli.cli_constants import CONFIGURATION_PATH
+
     if not CONFIGURATION_PATH.is_file():
         echo_error(
             "No configuration file found. Run 'dp init' to create it.",